+++ react_django/react-src/src/components/settings/studentSettings/studentSettings.css
